requirements.txt
streamlit
PyPDF2
crewai
langchain_community
python-dotenv
litellm
openai
